Fixed issue to edit/delete the image which is posted by some other user

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -92,13 +92,24 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model, HttpSession session) {
         Image image = imageService.getImage(imageId);
+        Integer userId=image.getUser().getId();
+        User loggedUser=(User)session.getAttribute("loggeduser");
+        if(loggedUser.getId()==userId){
+            String tags = convertTagsToString(image.getTags());
+            model.addAttribute("image", image);
+            model.addAttribute("tags", tags);
+            return "images/edit";
+        }
+        else{
+            model.addAttribute("image", image);
+            model.addAttribute("tags", image.getTags());
+            model.addAttribute("editError","Only the owner of the image can edit the image");
 
-        String tags = convertTagsToString(image.getTags());
-        model.addAttribute("image", image);
-        model.addAttribute("tags", tags);
-        return "images/edit";
+            return "/images/image";
+        }
+
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -140,9 +151,21 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, HttpSession session,Model model) {
+
+        Image image = imageService.getImage(imageId);
+        Integer userId=image.getUser().getId();
+        User loggedUser=(User)session.getAttribute("loggeduser");
+        if(loggedUser.getId()==userId){
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        }
+        else{
+            model.addAttribute("image", image);
+            model.addAttribute("deleteError","Only the owner of the image can delete the image");
+            return "/images/image";
+        }
+
     }
